[Kernel][Perf] Use architecture-family matmul configs
Move the tuned table and lookup into a dedicated module and select Hopper/Ada configs by CUDA device capability during batch-invariance initialization.
